modulo-pdv-final/modulo-pdv/backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
alembic==1.16.4
pydantic==2.11.7
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
fastapi-cors==0.0.6

